Android 9.0.0 Release 16 (PQ1A.181105.017.A1)

* tag 'android-9.0.0_r16': (34 commits)
  NuPlayer2CCDecoder: Add bound check before memcpy
  Fix race condition for cas sessions
  Get screenrecord to exclude black cutout
  Camera service: Extend UID check timeout to 300ms.
  Camera: Documentation update to remove physicalIds
  Fix information disclosure in mediadrmserver DO NOT MERGE
  Rework NuPlayer::getStats() synchronization
  Set rlimit rtprio for cameraserver
  SoftXaac: Handle error cases when decoder fails to initialize
  Camera: Documentation updates for calibration and distortion correction
  Check for overflow of crypto size
  audio policy: fix regression in Audio MMAP device selection
  Abort read() after CameraSource has stopped
  Camera: Handle abandoned surface in finishConfiguration
  Fix information disclosure in mediadrmserver
  stagefright: limit software renderer for actual OMX components
  stagefright: MediaCodec to allocate by owner
  Allow playing output to default output device when during uplink playback
  Camera: Mark stream as ABANDONED if dequeueBuffer returns DEAD_OBJECT
  media sync: forcefully limit correction by VideoFrameScheduler.
  ...

+++ media/libstagefright/codecs/mp3dec/src/pvmp3_framedecoder.cpp
@@ -299,7 +299,11 @@ ERROR_CODE pvmp3_framedecoder(tPVMP3DecoderExternal *pExt,
         }
 
 
-        bytes_to_discard = pVars->frame_start - pVars->sideInfo.main_data_begin - main_data_end;
+        // force signed computation; buffer sizes and offsets are all going to be
+        // well within the constraints of 32-bit signed math.
+        bytes_to_discard = pVars->frame_start
+                           - ((int32)pVars->sideInfo.main_data_begin)
+                           - ((int32)main_data_end);
 
 
         if (main_data_end > BUFSIZE)   /* check overflow on the buffer */
